pyproject.toml
[project]
name = "chat-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "n1ghtmare-dev",email = "oboev3691@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "aiogram (>=3.23.0,<4.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.127.0,<0.128.0)",
    "uvicorn (>=0.40.0,<0.41.0)"
]

[tool.poetry]
packages = [{include = "chat_tracker", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
